+++ matlab/changelog.md
@@ -3,6 +3,12 @@ IronClust, written by J. James Jun, Flatiron Institute, Simons Foundation
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.7.6] - 2019-6-14
+### Added
+- `post_merge_mode` 6 is added (default: 1)
+  - experimental feature. this will compute average waveforms per unit per drift-batch
+- `sort_mode` is reintroduced (default: 1)
+
 
 ## [4.7.5] - 2019-6-14
 ### Added